backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pypdf2==3.0.1
pdfplumber==0.10.3
python-dateutil==2.8.2

anthropic>=0.36.0
openai>=1.40.0
python-dotenv>=1.0.1
